requirements.txt
gradio==4.19.2
numpy==1.26.4
langgraph==0.0.40
google-generativeai==0.5.4
pyautogui

